refactor: modify pages folder path pages folder를 components 폴더와 동일한 계층에 위치시킴
